Add javadoc for 8 classes
