--- src/main/java/duke/ToDos.java
@@ -1,19 +1,0 @@
-package duke;
-
-import duke.Task;
-
-public class ToDos extends Task {
-
-    public ToDos(String description) {
-        super(description);
-    }
-
-    @Override
-    public String getSavingFormat() {
-        return "[T] | [" + getStatusIcon() + "] | " + description;
-    }
-    @Override
-    public String toString() {
-        return "[T][" + getStatusIcon() + "] " + description;
-    }
-}
