+++ src/main/java/duke/Ui.java
@@ -1,11 +1,21 @@
 package duke;
 
+/**
+ * Represents the User Interface (UI) of the application.
+ * Contains methods for user interactions.
+ */
 public class Ui {
 
+    /**
+     * Prints the greeting message to the user.
+     */
     public void sayhi() {
         System.out.println("Hello! I'm LimBay.\nWhat can I do for you?\n");
     }
 
+    /**
+     * Prints the goodbye message to the user.
+     */
     public void saybye() {
         System.out.println("Bye. Hope to see you again soon!");
     }
